+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Next Release
 * Improved: ``ImproperlyConfigured`` now raised if routes module is not defined
   in either ``init_app`` or in application configuration via
   ``VIA_ROUTES_MODULE``
+* Feature: Added support for ``VIA_ROUTES_NAME`` to set a common routes name
 
 2014.05.08
 ----------